Rename : message로 디렉토리 변경(#7)
