+++ backend/src/main/java/com/tripmarket/domain/chatting/repository/message/MessageRepository.java
@@ -1,7 +1,7 @@
-package com.tripmarket.domain.chatting.repository;
+package com.tripmarket.domain.chatting.repository.message;
 
 import com.tripmarket.domain.chatting.entity.Message;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface MessageRepository extends MongoRepository<Message, String> {
+public interface MessageRepository extends MongoRepository<Message, String>,CustomMessageRepository {
 }
