+++ backend/src/main/resources/static/chat-test.html
@@ -3,19 +3,24 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>1:1 채팅 테스트 (RabbitMQ)</title>
+    <title>1:1 채팅 테스트 (다수 사용자 지원)</title>
 </head>
 <body>
-<h1>1:1 채팅 테스트 (RabbitMQ)</h1>
+<h1>채팅 테스트 (다수 사용자 지원)</h1>
 
 <div>
     <label for="roomId">Room ID:</label>
-    <input type="text" id="roomId" value="1" placeholder="채팅방 ID 입력 (예: 1)" required>
+    <input type="text" id="roomId" value="1" placeholder="채팅방 ID 입력" required>
 </div>
 
 <div>
     <label for="userId">Your User ID:</label>
-    <input type="text" id="userId" value="1" placeholder="본인 User ID (1 or 2)" required>
+    <input type="text" id="userId" value="1" placeholder="본인 User ID" required>
+</div>
+
+<div>
+    <label for="targetUserId">Target User ID:</label>
+    <input type="text" id="targetUserId" placeholder="메시지 보낼 상대방 User ID" required>
 </div>
 
 <div>
@@ -35,7 +40,7 @@
 
     // WebSocket 연결 설정
     function connect() {
-        const socket = new SockJS('/chat');  // 서버의 WebSocket 엔드포인트
+        const socket = new SockJS('/chat');
         stompClient = Stomp.over(socket);
 
         stompClient.connect({}, function (frame) {
@@ -48,7 +53,7 @@
                 return;
             }
 
-            // 특정 채팅방의 메시지를 수신하는 경로 (/exchange/chat.exchange/chat.room.{roomId})
+            // 특정 채팅방의 메시지를 수신하는 경로
             stompClient.subscribe('/exchange/chat.exchange/chat.room.' + roomId, function (message) {
                 showMessage(JSON.parse(message.body));
             });
@@ -62,9 +67,10 @@
     function sendMessage() {
         const roomId = document.getElementById('roomId').value;
         const userId = document.getElementById('userId').value;
+        const targetUserId = document.getElementById('targetUserId').value;
         const messageContent = document.getElementById('message').value;
 
-        if (!roomId || !userId || !messageContent) {
+        if (!roomId || !userId || !targetUserId || !messageContent) {
             alert('모든 입력 값을 입력해주세요.');
             return;
         }
@@ -72,7 +78,7 @@
         const messageDto = {
             roomId: roomId,
             sender: userId,
-            toUserId: userId === "1" ? "2" : "1",  // 상대방 ID 설정
+            toUserId: targetUserId,
             content: messageContent
         };
